more wipping of new auth context in controllers
