+++ api/handlers/family/family_controller.go
@@ -4,10 +4,12 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/sebboness/yektaspoints/handlers"
 	"github.com/sebboness/yektaspoints/storage"
 )
 
 type FamilyController struct {
+	handlers.BaseController
 	familyDB storage.IFamilyStorage
 }
 
@@ -19,7 +21,15 @@ func NewFamilyController(ctx context.Context, env string) (*FamilyController, er
 		return nil, fmt.Errorf("failed to initialize family db: %w", err)
 	}
 
+	authContext, err := handlers.GetAuthContext()
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize auth context: %w", err)
+	}
+
 	return &FamilyController{
+		BaseController: handlers.BaseController{
+			AuthContext: authContext,
+		},
 		familyDB: familyDB,
 	}, nil
 }
