+++ api/handlers/points/handle_get_user_points.go
@@ -21,7 +21,7 @@ type getUserPointsHandlerResponse struct {
 
 func (c *PointsController) GetUserPointsHandler(cgin *gin.Context) {
 
-	// authInfo := handlers.GetAuthorizerInfo(cgin)
+	// authInfo := c.AuthContext.GetAuthorizerInfo(cgin)
 	// TODO: Implement a way to check that the requested user (a parent) has access to retrieve the points for the user
 	//       identified in this request via the user_id query parameter (their child).
