+++ api/handlers/points/points_controller.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/sebboness/yektaspoints/handlers"
 	"github.com/sebboness/yektaspoints/storage"
 	"github.com/sebboness/yektaspoints/util/log"
 )
@@ -11,6 +12,7 @@ import (
 var logger = log.Get()
 
 type PointsController struct {
+	handlers.BaseController
 	pointsDB storage.IPointsStorage
 	userDB   storage.IUserStorage
 }
@@ -23,7 +25,15 @@ func NewPointsController(ctx context.Context, env string) (*PointsController, er
 		return nil, fmt.Errorf("failed to initialize db: %w", err)
 	}
 
+	authContext, err := handlers.GetAuthContext()
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize auth context: %w", err)
+	}
+
 	return &PointsController{
+		BaseController: handlers.BaseController{
+			AuthContext: authContext,
+		},
 		pointsDB: db,
 		userDB:   db,
 	}, nil
