+++ api/handlers/userauth/user_auth_controller.go
@@ -4,10 +4,12 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/sebboness/yektaspoints/handlers"
 	"github.com/sebboness/yektaspoints/util/auth"
 )
 
 type UserAuthController struct {
+	handlers.BaseController
 	auth auth.AuthController
 }
 
@@ -17,7 +19,15 @@ func NewUserAuthController(ctx context.Context, env string) (*UserAuthController
 		return nil, fmt.Errorf("failed to initialize auth controller: %w", err)
 	}
 
+	authContext, err := handlers.GetAuthContext()
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize auth context: %w", err)
+	}
+
 	return &UserAuthController{
+		BaseController: handlers.BaseController{
+			AuthContext: authContext,
+		},
 		auth: authController,
 	}, nil
 }
